Cannot call sub to define a constant if the sub has not yet been defined, so just use a variable instead.
More refactoring to make flow of execution simpler (IMHO).

The test target works fine on my box without explicitly setting the
apache binary or apxs binary. It would be very bad form for a CPAN
module to die in the Makefile.PL! That'd break any attempt to install
it via a CPAN{PLUS} shell. I'll figure out how to make this optional
in a future commit.


git-svn-id: https://svn.apache.org/repos/asf/perl/Apache-SizeLimit/trunk@419247 13f79535-47bb-0310-9956-ffa450edef68